end place order

+++ lib/Feather/Home/data/model/proudact_model.dart
@@ -1,6 +1,6 @@
 class ProudactModel {
   final String name;
-  final double price;
+  final int price;
   final String des;
   final String image;
   ProudactModel({
